requirements.txt
opencv-python==4.8.0.74
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
tk==0.1.0
flask==2.3.3
tkcalendar==1.6.1
